requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
python-dotenv==1.0.0
elevenlabs==1.50.3
groq==0.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiohttp==3.9.1
pyaudio==0.2.14
pandas==2.1.4
plotly==5.18.0
pydantic>=2.4.2,<3.0.0
pydantic-core>=2.18.2,<3.0.0
email-validator>=2.1.0
openai==1.10.0